Move verify functions from main to license-verify
